+++ src/metriclearn/neighborcompanalysis.h
@@ -11,7 +11,7 @@ namespace segm {
     public:
 
         NeighborCompAnalysis(Matrix<float> &data, Vector<int> &label, int output_dim, int _iterations = 10000,
-                             float _learn_rate = 1e-2, bool _verbose = false);
+                             double _learn_rate = 1e-2, bool _verbose = false);
 
         Matrix<float> transform(const Matrix<float> &data) const;
         Matrix<double> transform(const Matrix<double> &data) const;
@@ -28,7 +28,7 @@ namespace segm {
         int d_in;       // input dimension
         int d_out;      // output dimension
 
-        float learn_rate = 1e-2;
+        double learn_rate = 1e-2;
         int iterations = 10000;
         bool verbose = false;
         bool executed = false;
